Extract FeedLoader test helpers into a shared protocol extension to…
… remove duplication
